pyproject.toml
[project]
name = "moneymonkey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.26.0",
    "aiosqlite>=0.22.1",
    "cryptography>=46.0.5",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.3.0",
    "gradio>=6.10.0",
    "gradio-client>=2.4.0",
    "langchain>=1.2.14",
    "langchain-gigachat>=0.5.0",
    "langchain-openai>=1.1.12",
    "openai>=2.30.0",
    "pandas>=3.0.2",
    "pandasql",
    "python-dotenv>=1.2.2",
]
